+++ page-templates/partials/content-page.php
@@ -18,12 +18,24 @@
 	<div class="entry-content" itemprop="mainContentOfPage">
 
 		<?php the_content(); ?>
-		<?php
-			wp_link_pages( array(
-				'before' => '<div class="page-links">' . __( 'Pages:', 'digistarter' ),
-				'after'  => '</div>',
-			) );
-		?>
+
+		<?php if ( function_exists( 'get_the_post_navigation' ) ) {
+
+				echo get_the_post_navigation( array(
+					'prev_text'	=> __('&larr; Previous Page', 'digistarter'),
+					'next_text'	=> __( 'Next Page &rarr;', 'digistarter'),
+					'screen_reader_text' => __( 'Page navigation', 'digistarter' )
+				));
+
+			} else {
+
+				wp_link_pages( array(
+					'before' => '<div class="page-links">' . __( 'Pages:', 'digistarter' ),
+					'after'  => '</div>',
+					) );
+
+			} ?>
+
 
 	</div><!-- .entry-content -->
 	<?php edit_post_link( __( 'Edit', 'digistarter' ), '<footer class="entry-meta"><span class="edit-link">', '</span></footer>' ); ?>