+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# Changelog
+### 1.1.10
+-------------
+
+**Features:**
+* Added support for get_the_archive_title() in the archive template
+* Added support for 
+
+### 1.1.10
+-------------
+
+**Features:**
+* General tweaks to Customizer Settings Area. Utilizing panels to separate settings.
+
+**Bugfixes:**
+*
+
 ### 1.1.9
 -------------